refactor(FullyAutoAndSemiAutoTools): 重构UID识别模块并更新依赖版本
- 将uid.js重命名为tool.js并导出相关函数
- 移除旧的OCR UID识别逻辑，改用genshin.uid()方法
- 更新manifest.json中的版本号从0.0.1到0.0.2
- 升级bgi_version从0.54.3到0.61.0
- 修改main.js中的导入语句和UID获取方式
- 在initRecord函数中先调用toMainUi()确保在主界面
- 更新README.md中的文件结构说明
