+++ repo/js/FullyAutoAndSemiAutoTools/manifest.json
@@ -1,8 +1,8 @@
 {
   "manifest_version": 1,
   "name": "全自动或半自动工具箱",
-  "version": "0.0.1",
-  "bgi_version": "0.54.3",
+  "version": "0.0.2",
+  "bgi_version": "0.61.0",
   "key": "PGCSBY37NJA",
   "description": "",
   "saved_files": [],
